stack.yaml: Make Stack builds work again using the latest snapshot
Completely replace the old, broken Stack configuration with a working
build using the current latest Stackage lts-23.27 for ghc-9.8.4.

This was previously using snapshot lts-3.5 for ghc-7.10.2, which is
more than ten years old, and this build has probably been broken for a
long time. (Very old Stackage snapshots tend not to work with modern
versions of `stack`, even if the code itself is still compatible.)