.gitignore: Ignore /stack.yaml.lock and root /.stack-work/
This also removes the existing committed stack.yaml.lock.

stack.yaml.lock is normally committed to lock down an exact set of
dependencies for the build. However, in this project we're generally
preferring to regularly recreate the build plan with the latest
versions of dependencies in order to keep up to date with them.

Stack still locks down a fair amount even when generating a new
stack.yaml.lock, but developers on this project don't use Stack; the
configuration exists mainly to make it quick and easy to do a build
for those who don't want to set up their own GHC and Cabal toolchain.

A stack.yaml.lock or .stack-work/ appearing in any directory other
than the root of the project indicates a problem with the build,
so we do not ignore those in subdirectories.

(There are probably other things here that also should be fixed to
the root of the project, but we can deal with that in another commit.)