stack.yaml: Add tasty-hunit dependency for tests
`stack build --test` (which runs a lot more than what .gh/wf/haskell
workflow does) takes a half hour or more and mostly passes, but the
test-nebula-plugin suite fails. This can be reproduced with:

    stack build --test --ta '-p /add_assoc/'

The issue appears to be that tests_nebula_plugin/Main.hs wants to run
`cabal clean`, but the `cabal` executable doesn't exist in a pure
Stack environment. (Stack uses the cabal library to do the builds
itself, rather than calling the `cabal` command-line program.)

Nonetheless, this commit is still forward progress. If fixing
tests_nebula_plugin/Main.hs not to call the `cabal` executable
is difficult, it may be reasonable not to support running those
tests with Stack.