Fix Stack Builds (#331)
* .gitignore: Ignore /stack.yaml.lock and root /.stack-work/

This also removes the existing committed stack.yaml.lock.

stack.yaml.lock is normally committed to lock down an exact set of
dependencies for the build. However, in this project we're generally
preferring to regularly recreate the build plan with the latest
versions of dependencies in order to keep up to date with them.

Stack still locks down a fair amount even when generating a new
stack.yaml.lock, but developers on this project don't use Stack; the
configuration exists mainly to make it quick and easy to do a build
for those who don't want to set up their own GHC and Cabal toolchain.

A stack.yaml.lock or .stack-work/ appearing in any directory other
than the root of the project indicates a problem with the build,
so we do not ignore those in subdirectories.

(There are probably other things here that also should be fixed to
the root of the project, but we can deal with that in another commit.)

* stack.yaml: Make Stack builds work again using the latest snapshot

Completely replace the old, broken Stack configuration with a working
build using the current latest Stackage lts-23.27 for ghc-9.8.4.

This was previously using snapshot lts-3.5 for ghc-7.10.2, which is
more than ten years old, and this build has probably been broken for a
long time. (Very old Stackage snapshots tend not to work with modern
versions of `stack`, even if the code itself is still compatible.)

* stack.yaml: Add tasty-hunit dependency for tests

`stack build --test` (which runs a lot more than what .gh/wf/haskell
workflow does) takes a half hour or more and mostly passes, but the
test-nebula-plugin suite fails. This can be reproduced with:

    stack build --test --ta '-p /add_assoc/'

The issue appears to be that tests_nebula_plugin/Main.hs wants to run
`cabal clean`, but the `cabal` executable doesn't exist in a pure
Stack environment. (Stack uses the cabal library to do the builds
itself, rather than calling the `cabal` command-line program.)

Nonetheless, this commit is still forward progress. If fixing
tests_nebula_plugin/Main.hs not to call the `cabal` executable
is difficult, it may be reasonable not to support running those
tests with Stack.